Show or hide stream depending on shared secret mode & shared secret code
The SessionBuilder takes the boolean sessionMode(preferences) to tell the Session whether it should add the string "a=shared-secret:hashHere" into the stream's description.
Then, RTSPServerWorker will parse it and build the corresponding ReceiveSession. After that, the receivingsession will be used to build a new Streaming object and add it into mServerSessions, a list which will be used to update the GUI's stream list. I just have to check whether the Streaming is a "shared secret" (a boolean that shows the result after matching in serverworker) and show it or not accordingly.

Added new class "Sha256" to hash and check hashes
Added a switch to root_preferences to activate shared secret mode: Later will add a edittext to set the shared secret code